# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NextDocs API",
    "description": "API documentation for NextDocs platform",
    "version": "1.0.0",
    "contact": {
      "name": "NextDocs Support",
      "url": "https://nextdocs.example.com"
    }
  },
  "servers": [
    {
      "url": "https://api.nextdocs.example.com/v1",
      "description": "Production API"
    },
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local Development"
    }
  ],
  "paths": {
...
    "/search": {
      "get": {
        "summary": "Search documentation",
        "description": "Search across all documentation content",
        "operationId": "searchDocs",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "description": "Language locale",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["en", "es", "fr", "de"],
              "default": "en"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SearchResult"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "SearchResult": {
        "type": "object",
        "required": ["slug", "title"],
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "locale": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
...
    }
  }
}